Broke up codecs
